workin on repositories

+++ backend/tattooManager/src/main/kotlin/com/tattoomanager/tattooManager/port/repository/BudgetRepository.kt
@@ -4,5 +4,5 @@ import com.tattoomanager.tattooManager.domain.Budget
 
 interface BudgetRepository {
     fun save(budget: Budget): Budget
-    fun findByUserId(userId: Int): List<Budget>
+    fun findByUserId(userId: Long): List<Budget>
 }
